Create Running man
On the first day, the athlete ran X kilometers. On each subsequent day, he increased the mileage by 15% from the previous day. You need to determine the number of the day on which the athlete's mileage was at least Y kilometers. The number Y itself will be input to the program.