Create Traversing matrix elements - 1
An integer square matrix of size N x N is given. It is necessary to bypass the elements of this matrix from top to bottom from left to right and display the elements in this order in the form of a table. 
The program takes as input a natural number N – the number of rows and columns of the matrix. In each of the next N lines, N integers are written – elements of the matrix. All numbers in the input data do not exceed 100 in absolute value.