Create Traversing matrix elements - 2
An integer square matrix of size N x N is given. It is necessary to bypass the elements of this matrix from bottom to top from right to left and display the elements in this order in the form of a table. 
The program takes as input a natural number N – the number of rows and columns of the matrix. In each of the following N lines, N integers are written – elements of the matrix.