Create Traversing matrix elements - 3
An integer matrix consisting of N rows and M columns is given. It is necessary to bypass the elements of this matrix from right to left from top to bottom and display the elements in this order in the form of a table. 
The program accepts two natural numbers N and M as input – the number of rows and columns of the matrix. In each of the following N lines, M integers are written – elements of the matrix.