Create Beautiful matrix
Here is a 5 ×5 matrix consisting of 24-x zeros and a single one. The rows of the matrix are numbered with numbers from 1 to 5 from top to bottom, the columns of the matrix are numbered with numbers from 1 to 5 from left to right. In one move , it is allowed to apply one of the following two transformations to the matrix:

Swap two adjacent rows of the matrix, that is, rows with numbers i and i + 1 for some integer i (1 ≤ i < 5).
Swap two adjacent columns of the matrix, that is, columns with numbers j and j + 1 for some integer j (1 ≤ j < 5).
You think that the matrix will look beautiful if the only unit of this matrix is in its center (in the cell that is at the intersection of the third row and the third column). Calculate the minimum number of moves it will take to make the matrix beautiful.

Input data

The input data consists of five rows, each of which contains five integers: the j-th number in the i-th row of the input data denotes an element of the matrix standing at the intersection of the i-th row and the j-th column. It is guaranteed that the matrix consists of 24-x zeros and a single one.

Output data

Print a single integer — the minimum number of actions required to make the matrix beautiful.