Create The sum of rows and columns of a two-dimensional array
An integer two-dimensional array consisting of N rows and M columns is given. It is required to calculate the sum of the elements in each row and in each column.
The program receives two natural numbers N and M as input – the number of rows and columns of a two-dimensional array. In each of the following N lines, M integers are written – the elements of the array. All numbers in the input data do not exceed 1000 in absolute value.
In the first line, you need to output N numbers – the sum of the array elements for each row separately.
In the second row, in a similar format, print M numbers – the sum of elements for each column.