Create Dict
Your program receives a string as input, you need to count how many times each letter in this line has been encountered case-insensitive, while numbers and punctuation characters need to be skipped. The answer must be stored in a dictionary in which the key is a letter, and the value is the number of times this letter has occurred in the string. As an answer, you need to output a dictionary