Create *args
Let's now create a print_goods function that prints a shopping list. It will accept an arbitrary number of values as input, and we will consider any non-empty strings as a product. That is, numbers, lists, dictionaries and other non-string objects you will need to ignore. The print_goods function should print a list of products in the form: <Product serial number>. <Product name> (see the example below). If no goods are found in the transmitted values, it is necessary to print the text "No goods"