Create recursion_2
Write a list_sum_recursive function that accepts a list of integers as input and returns the sum of the elements of the passed list. Do not forget that you need to implement this using recursion. 
Your task is only to write the definition of the list_sum_recursive function