Create Linear search
Linear search, also known as sequential search, this method is used to search for an item in a list. Linear search is one of the basic algorithms that you should become familiar with while studying programming. The essence of the algorithm is as follows: you have to check each element of the list sequentially one by one until you find the element you are interested in or until the entire list ends.

Input data

The program receives as input in one line the elements of the list - integers separated by a space. The number of elements is arbitrary

And on the next line, a single number r is entered - the search value

Output data

Your task is to implement a linear algorithm for searching for the entered value of r. If successful, print the ordinal number (index) of the first found item in the list, provided that indexing starts with one. If this element is missing, it is necessary to output the Error Value string